Let auto-provision call method instead of going through a grpc call
